Events don't need `CallEventsIngestionOperation` instantiation/pool anymore
Needs tidyup but we now run coroutines needed for events on a single MonoBehaviour.
Moved some more code from SdkOperations to BackendOperations
